+++ Events/OnAnimationEnd.ts
@@ -16,9 +16,19 @@ import { animationendEventName } from '../Tools/PrefixedProperties';
  *
  * animationEnd.off()
  *
+ * // To watch for a animation end:
+ * onAnimationEnd( $elem, {
+ * "animationName": "name-of-my-css-animation"
+ * } )
+ *
+ * // To watch a animation end on a pseudo element like "::after":
+ * onAnimationEnd( $elem, {
+ * "pseudoElement": "after"
+ * } )
+ *
  * @returns Return a standard Promise + an .off() function to cancel event
  */
-export default function onAnimationEnd( $element: Element ): Promise<any> & { off(); } {
+export default function onAnimationEnd( $element: Element, options: { pseudoElement?: 'after' | 'before' | 'both', animationName?: string } = {} ): Promise<any> & { off(); } {
     let _resolve;
 
     const promise = new Promise( function( resolve ) {
@@ -32,7 +42,14 @@ export default function onAnimationEnd( $element: Element ): Promise<any> & { of
 
 
     function onAnimationEnd( e ) {
-        if ( e.target !== $element ) {
+        if (
+            e.target !== $element ||
+            ( options.animationName && options.animationName !== e.animationName ) ||
+            ( options.pseudoElement === 'after' && e.pseudoElement !== '::after' ) ||
+            ( options.pseudoElement === 'before' && e.pseudoElement !== '::before' ) ||
+            ( options.pseudoElement === 'both' && !e.pseudoElement ) ||
+            ( !options.pseudoElement && e.pseudoElement !== '' )
+        ) {
             return;
         }